재활용 오브젝트 초기화 오류 수정

+++ Assets/Scripts/Interface/IHealth.cs
@@ -9,12 +9,12 @@ public interface IHealth
     /// <summary>
     /// 현재 체력
     /// </summary>
-    public int Health { get; set; }
+    public float Health { get; set; }
 
     /// <summary>
     /// 최대 체력
     /// </summary>
-    public int MaxHealth { get; set; }
+    public float MaxHealth { get; set; }
 
     /// <summary>
     /// 피격 시 실행되는 함수
